fix: more css standardization, consistent ui/ux for transformations, kebab renaming
