+++ libs/renderer/src/components/cell-defaults/collapse-transformation-cell/index.ts
@@ -1,2 +1,2 @@
-export * from "./CollapseTransformationCell";
+export * from "./collapse-transformation-cell";
 export * from "./config";
