+++ libs/renderer/src/components/cell-defaults/column-type-transformation-cell/index.ts
@@ -1,2 +1,2 @@
-export * from "./ColumnTypeTransformationCell";
+export * from "./column-type-transformation-cell";
 export * from "./config";
