+++ libs/renderer/src/components/cell-defaults/cumulative-sum-transformation-cell/index.ts
@@ -1,2 +1,2 @@
-export * from "./CumulativeSumTransformationCell";
 export * from "./config";
+export * from "./cumulative-sum-transformation-cell";
