+++ libs/renderer/src/components/cell-defaults/encode-column-transformation-cell/index.ts
@@ -1,2 +1,2 @@
 export * from "./config";
-export * from "./EncodeColumnTransformationCell";
+export * from "./encode-column-transformation-cell";
