+++ libs/renderer/src/components/cell-defaults/shared/transformation/column-transformation-field.tsx
@@ -125,7 +125,10 @@ export const ColumnTransformationField: ColumnTransformationFieldComponent =
 							)}
 						</Button>
 					</PopoverTrigger>
-					<PopoverContent className="w-[--radix-popover-trigger-width] p-0">
+					<PopoverContent
+						align="start"
+						className="w-[var(--radix-popover-trigger-width)] p-0"
+					>
 						<Command>
 							<CommandInput placeholder="Search columns..." />
 							<CommandEmpty>No columns found.</CommandEmpty>
@@ -175,7 +178,10 @@ export const ColumnTransformationField: ColumnTransformationFieldComponent =
 						)}
 					</Button>
 				</PopoverTrigger>
-				<PopoverContent className="w-[--radix-popover-trigger-width] p-0">
+				<PopoverContent
+					align="start"
+					className="w-[var(--radix-popover-trigger-width)] p-0"
+				>
 					<Command>
 						<CommandInput placeholder="Search columns..." />
 						<CommandEmpty>No columns found.</CommandEmpty>
