+++ libs/renderer/src/components/cell-defaults/shared/transformation/index.ts
@@ -1,8 +1,7 @@
-export { ColumnTransformationField } from "./ColumnTransformationField";
-export { EncodeColumnCheckboxTransformationField } from "./EncodeColumnCheckboxTransformationField";
-export { MultiCellColumnTransformationField } from "./MultiCellColumnTransformationField";
-export { TransformationCellInput } from "./TransformationCellInput";
-export { TransformationCellRunActionButton } from "./TransformationCellRunActionButton";
-export { TransformationMultiCellInput } from "./TransformationMultiCellInput";
+export { ColumnTransformationField } from "./column-transformation-field";
+export { MultiCellColumnTransformationField } from "./multi-cell-column-transformation-field";
 export * from "./transformation.constants";
 export * from "./transformation.types";
+export { TransformationCellInput } from "./transformation-cell-input";
+export { TransformationCellRunActionButton } from "./transformation-cell-run-action-button";
+export { TransformationMultiCellInput } from "./transformation-multi-cell-input";
