+++ libs/renderer/src/components/cell-defaults/shared/transformation/transformation-cell-input.tsx
@@ -61,12 +61,18 @@ export const TransformationCellInput: TransformationCellInputComponent = (
 					});
 				}}
 			>
-				<SelectTrigger className="h-[30px] w-[200px]">
-					<Network className="mr-1 size-4 shrink-0" />
-					<SelectValue placeholder="Select frame" />
+				<SelectTrigger className="h-9 w-[200px] min-w-0">
+					<div className="flex min-w-0 flex-1 items-center gap-1">
+						<Network className="size-4 shrink-0" />
+						<SelectValue placeholder="Select frame" />
+					</div>
 				</SelectTrigger>
 				<SelectContent>
-					{frame?.options.map((c) => (
+					{Array.from(
+						new Map(
+							(frame?.options ?? []).map((c) => [c.id, c]),
+						).values(),
+					).map((c) => (
 						<SelectItem key={c.id} value={c.id}>
 							{c.parameters.frameVariableName}
 						</SelectItem>
