+++ libs/renderer/src/components/cell-defaults/timestamp-transformation-cell/index.ts
@@ -1,2 +1,2 @@
 export * from "./config";
-export * from "./TimestampTransformationCell";
+export * from "./timestamp-transformation-cell";
