+++ libs/renderer/src/components/cell-defaults/update-row-transformation-cell/index.ts
@@ -1,2 +1,2 @@
 export * from "./config";
-export * from "./UpdateRowTransformationCell";
+export * from "./update-row-transformation-cell";
