+++ libs/renderer/src/components/cell-defaults/update-row-transformation-cell/update-row-transformation-cell.tsx
@@ -56,15 +56,9 @@ export const UpdateRowTransformationCell: CellComponent<UpdateRowTransformationC
 		const { state } = useBlocks();
 
 		const targetCell: CellState<QueryImportCellDef> = computed(() => {
-			let c: CellState<QueryImportCellDef> | undefined;
-			Object.values(state.queries).forEach((query) => {
-				if (query.cells[cell.parameters.targetCell.id]) {
-					c = query.cells[
-						cell.parameters.targetCell.id
-					] as CellState<QueryImportCellDef>;
-				}
-			});
-			return c;
+			return cell.query.cells[
+				cell.parameters.targetCell.id
+			] as CellState<QueryImportCellDef>;
 		}).get();
 
 		const cellTransformation: Transformation<UpdateRowTransformationDef> =
@@ -80,19 +74,10 @@ export const UpdateRowTransformationCell: CellComponent<UpdateRowTransformationC
 		}).get();
 
 		const frames = useMemo(() => {
-			const frameList = [];
-			Object.keys(state.queries).forEach((queryKey) => {
-				const query = state.queries[queryKey];
-				Object.values(query.cells).forEach((cell) => {
-					if (
-						cell.widget === "query-import" ||
-						cell.widget === "data-import"
-					) {
-						frameList.push(cell);
-					}
-				});
-			});
-			return frameList;
+			return Object.values(cell.query.cells).filter(
+				(c) =>
+					c.widget === "query-import" || c.widget === "data-import",
+			);
 		}, []);
 
 		const getTextFieldType = (dataType: string): string => {
@@ -127,9 +112,8 @@ export const UpdateRowTransformationCell: CellComponent<UpdateRowTransformationC
 			});
 
 		if (
-			(!doesFrameExist &&
-				!cellTransformation.parameters.compareColumn.name) ||
-			!targetCell.isExecuted
+			!doesFrameExist &&
+			!cellTransformation.parameters.compareColumn.name
 		) {
 			return (
 				<TransformationCellInput
