+++ libs/renderer/src/components/cell-defaults/uppercase-transformation-cell/index.ts
@@ -1,2 +1,2 @@
 export * from "./config";
-export * from "./UppercaseTransformationCell";
+export * from "./uppercase-transformation-cell";
