+++ libs/renderer/src/components/cell-defaults/uppercase-transformation-cell/uppercase-transformation-cell.tsx
@@ -10,7 +10,7 @@ import {
 import type { QueryImportCellDef } from "../query-import-cell";
 import {
 	type ColumnInfo,
-	EncodeColumnCheckboxTransformationField,
+	ColumnTransformationField,
 	type Transformation,
 	type TransformationCellDef,
 	TransformationCellInput,
@@ -19,44 +19,38 @@ import {
 	type TransformationTargetCell,
 } from "../shared";
 
-export interface EncodeColumnTransformationDef
-	extends TransformationDef<"encode-column"> {
-	key: "encode-column";
+export interface UppercaseTransformationDef
+	extends TransformationDef<"uppercase"> {
+	key: "uppercase";
 	parameters: {
 		columns: ColumnInfo[];
 	};
 }
 
-export interface EncodeColumnTransformationCellDef
-	extends TransformationCellDef<"encode-column-transformation"> {
-	widget: "encode-column-transformation";
+export interface UppercaseTransformationCellDef
+	extends TransformationCellDef<"uppercase-transformation"> {
+	widget: "uppercase-transformation";
 	parameters: {
-		transformation: Transformation<EncodeColumnTransformationDef>;
+		transformation: Transformation<UppercaseTransformationDef>;
 		targetCell: TransformationTargetCell;
 	};
 }
 
-export const EncodeColumnTransformationCell: CellComponent<EncodeColumnTransformationCellDef> =
+export const UppercaseTransformationCell: CellComponent<UppercaseTransformationCellDef> =
 	observer((props) => {
 		const { cell, isExpanded } = props;
 		const { state } = useBlocks();
 
 		const targetCell: CellState<QueryImportCellDef> = computed(() => {
-			let c: CellState<QueryImportCellDef> | undefined;
-			Object.values(state.queries).forEach((query) => {
-				if (query.cells[cell.parameters.targetCell.id]) {
-					c = query.cells[
-						cell.parameters.targetCell.id
-					] as CellState<QueryImportCellDef>;
-				}
-			});
-			return c;
+			return cell.query.cells[
+				cell.parameters.targetCell.id
+			] as CellState<QueryImportCellDef>;
 		}).get();
 
-		const cellTransformation: Transformation<EncodeColumnTransformationDef> =
+		const cellTransformation: Transformation<UppercaseTransformationDef> =
 			computed(() => {
 				return cell.parameters
-					.transformation as Transformation<EncodeColumnTransformationDef>;
+					.transformation as Transformation<UppercaseTransformationDef>;
 			}).get();
 
 		const doesFrameExist: boolean = computed(() => {
@@ -66,29 +60,17 @@ export const EncodeColumnTransformationCell: CellComponent<EncodeColumnTransform
 		}).get();
 
 		const frames = useMemo(() => {
-			const frameList = [];
-			Object.keys(state.queries).forEach((queryKey) => {
-				const query = state.queries[queryKey];
-				Object.values(query.cells).forEach((cell) => {
-					if (
-						cell.widget === "query-import" ||
-						cell.widget === "data-import"
-					)
-						frameList.push(cell);
-				});
-			});
-			return frameList;
+			return Object.values(cell.query.cells).filter(
+				(c) =>
+					c.widget === "query-import" || c.widget === "data-import",
+			);
 		}, []);
 
 		const helpText = cell.parameters.targetCell.id
 			? `Run Cell ${cell.parameters.targetCell.id} to define the target frame variable before applying a transformation.`
 			: "A Python or R target frame variable must be defined in order to apply a transformation.";
 
-		if (
-			(!doesFrameExist &&
-				!cellTransformation.parameters.columns.length) ||
-			!targetCell.isExecuted
-		) {
+		if (!doesFrameExist && !cellTransformation.parameters.columns.length) {
 			return (
 				<TransformationCellInput
 					isExpanded={isExpanded}
@@ -115,12 +97,17 @@ export const EncodeColumnTransformationCell: CellComponent<EncodeColumnTransform
 						{!doesFrameExist ? (
 							<em>{helpText}</em>
 						) : (
-							"Obfuscate the values of a column"
+							"Change the values of the selected columns to uppercase"
 						)}
 					</span>
-					<EncodeColumnCheckboxTransformationField
+					<ColumnTransformationField
 						disabled={!doesFrameExist}
 						cell={cell}
+						selectedColumns={
+							cellTransformation.parameters.columns ?? []
+						}
+						multiple
+						columnTypes={["STRING"]}
 						onChange={(newColumns: ColumnInfo[]) => {
 							state.dispatch({
 								message: ActionMessages.UPDATE_CELL,
